Move trigno into device_managers
