+++ bomi/device_managers/trigno/__init__.py
