+++ bomi/device_managers/trigno/avanti_modes.tsv
