+++ bomi/device_managers/trigno/client.py
