+++ bomi/device_managers/trigno/datastructure.py
