+++ bomi/widgets/start_react_widget.py
@@ -21,7 +21,7 @@ from bomi.widgets.window_mixin import WindowMixin
 from bomi.audio.player import TonePlayer, AudioCalibrationWidget
 import bomi.colors as bcolors
 
-from trigno_sdk.client import TrignoClient
+from bomi.device_managers.trigno.client import TrignoClient
 
 
 def _print(*args):
